% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{approxFpop}
\alias{approxFpop}
\title{approxFpop}
\usage{
approxFpop(
  data,
  penalty,
  approximation = "rectangle",
  intersection = "all",
  exclusion = "all",
  NbOfCands = FALSE,
  NbOfExclus = FALSE
)
}
\arguments{
\item{data}{is a matrix of data (p-rows x n-columns).}

\item{penalty}{is a value of penalty (a non-negative real number).}

\item{approximation}{is the type of approximation : 'rectangle' or 'sphere' (by default, 'rectangle').}

\item{intersection}{is the type of intersection : 'empty', 'all', 'last' or 'random' (by default, 'last').}

\item{exclusion}{is the type of intersection : 'empty', 'all'or 'random'(by default, 'all').
The following parameter combinations are implemented:
(approximation = 'rectangle', intersection ='empty', exclusion ='empty') (PELT-method)
(approximation = 'rectangle', intersection ='all', exclusion ='all')
(approximation = 'rectangle', intersection ='all', exclusion ='empty')
(approximation = 'rectangle', intersection ='empty', exclusion ='all')
(approximation = 'rectangle', intersection ='last', exclusion ='all')
(approximation = 'rectangle', intersection ='last', exclusion ='random')
(approximation = 'rectangle', intersection ='all', exclusion ='random')
(approximation = 'rectangle', intersection ='random', exclusion ='random')
(approximation = 'sphere', intersection ='last', exclusion ='all')
(approximation = 'sphere', intersection ='last', exclusion ='random')
(approximation = 'sphere', intersection ='random', exclusion ='random')}

\item{NbOfCands}{is the logical parameter (if NbOfCands = TRUE, than the file "NbOfCands.txt" contains the number of change candidates for each iteration.}

\item{NbOfExclus}{is the logical parameter (if NbOfExclus = TRUE, than the file "NbOfExclus.txt" contains the label of candidate and the number of exclusion for change candidates for each iteration.}
}
\value{
a list of  elements  = (changes, means, UnpenalizedCost).

\describe{
\item{\code{changes}}{is the changepoint vector that gives the last index of each segment for the p-variate time series.}
\item{\code{means}}{is the list of successive means for the p-variate time series.}
\item{\code{UnpenalizedCost}}{is a number equal to the global cost.}
}
}
\description{
FPOP method (using the rectangle approximation of the sets) for  the multiple changepoint detection.
}
\examples{
N <- 100
Chpt <-50
Means <-  matrix(c(0,1,1,10), nrow = 2)
Noise <- 1
Dim <- 2
Penality <- 2*Dim*log(N)
time_series <- rnormChanges(p = Dim, n = N, changes = Chpt, means = Means, noise = Noise)
time_series <- rnormChanges(p = 2, n = N, changes = NULL, means = matrix(0, ncol = 1, nrow = 2), noise = 1)
Approx <- list()
Approx[[1]] <- approxFpop(data = time_series, penalty = Penality, approximation = 'rectangle', intersection = 'all', exclusion = 'all', NbOfCands = FALSE, NbOfExclus = FALSE)
Approx[[2]] <- approxFpop(data = time_series, penalty = Penality, approximation = 'rectangle', intersection = 'all', exclusion = 'empty', NbOfCands = FALSE, NbOfExclus = FALSE)
Approx[[3]] <- approxFpop(data = time_series, penalty = Penality, approximation = 'rectangle', intersection = 'all', exclusion = 'random', NbOfCands = FALSE, NbOfExclus = FALSE)
Approx[[4]] <- approxFpop(data = time_series, penalty = Penality, approximation = 'rectangle', intersection = 'empty', exclusion = 'all', NbOfCands = FALSE, NbOfExclus = FALSE)
Approx[[5]] <- approxFpop(data = time_series, penalty = Penality, approximation = 'rectangle', intersection = 'last', exclusion = 'all', NbOfCands = FALSE, NbOfExclus = FALSE)
Approx[[6]] <- approxFpop(data = time_series, penalty = Penality, approximation = 'rectangle', intersection = 'last', exclusion = 'random', NbOfCands = FALSE, NbOfExclus = FALSE)
Approx[[7]] <- approxFpop(data = time_series, penalty = Penality, approximation = 'rectangle',intersection = 'random', exclusion = 'all', NbOfCands = FALSE, NbOfExclus = FALSE)
Approx[[8]] <- approxFpop(data = time_series, penalty = Penality, approximation = 'rectangle', intersection = 'random', exclusion = 'random', NbOfCands = FALSE, NbOfExclus = FALSE)
Approx[[9]] <- approxFpop(data = time_series, penalty = Penality, approximation = 'sphere', intersection = 'last', exclusion = 'all', NbOfCands = FALSE, NbOfExclus = FALSE)
Approx[[10]] <- approxFpop(data = time_series, penalty = Penality, approximation = 'sphere', intersection = 'random', exclusion = 'all', NbOfCands = FALSE, NbOfExclus = FALSE)
Approx[[11]] <- approxFpop(data = time_series, penalty = Penality, approximation = 'sphere', intersection = 'random', exclusion = 'random', NbOfCands = FALSE, NbOfExclus = FALSE)
Approx[[12]] <- approxFpop(data = time_series, penalty = Penality, approximation = 'sphere', intersection = 'last', exclusion = 'random', NbOfCands = FALSE, NbOfExclus = FALSE)

Approx[[13]] <- approxFpop(data = time_series, penalty = Penality, approximation = 'rectangle', intersection = 'empty', exclusion = 'empty', NbOfCands = FALSE, NbOfExclus = FALSE)
Approx
}
