% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPOPapproximation.R
\name{rnormChanges}
\alias{rnormChanges}
\title{rnormChanges}
\usage{
rnormChanges(
  p,
  n,
  changes = NULL,
  means = matrix(0, ncol = 1, nrow = p),
  noise = 1
)
}
\arguments{
\item{p}{parameter of dimension.}

\item{n}{number of data point.}

\item{changes}{a vector of increasing changepoint indices (last element is always less then n).
By default, 'changes = NULL' (the data without changepoints).}

\item{means}{matrix of successive means for data, by default 'means = matrix(0, ncol = 1, nrow = p)'.}

\item{noise}{standard deviation of an additional normal noise, by default 'noise = 1'.}
}
\value{
matrix of data of dimension p x n with a given values of means by the segmentation.
}
\description{
Generation of data (normal distribution) of dimension p with a given values of means and changes
}
\examples{
N <- 10
Chpt <-5
Means <-  matrix(c(0,1,1,10), nrow = 2)
Noise <- 1
Dim <- 2
Penality <- 2*Dim*log(N)
time_series1 <- rnormChanges(p = Dim, n = N, changes = Chpt, means = Means, noise = Noise)
time_series2 <- rnormChanges(p = Dim, n = N, changes = NULL, means = matrix(0, ncol = 1, nrow = Dim), noise = Noise)
}
